+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ##TBD
+* Make webview parameters optional in MSALSignoutParameters #1086
+* Support wiping external account #1085
+* Normalize account ID for cache lookups (#1084)
+* Add documentation for Proof-of-Possession for Access tokens.
 * Support forgetting cached account (#1077)
 * Add SSO Seeding call in MSAL Test MacApp 
 * Fix custom webview bug in MSAL Test MacApp